fix: binary data not showing in dashboard chat.
