+++ src/components/panels/chat/ChatMessages.jsx
@@ -7,88 +7,14 @@ import {
     VStack,
     Spinner,
     Image,
-    Button,
-    Link,
     useColorMode,
 } from "@chakra-ui/react";
-import { ExternalLinkIcon, DownloadIcon } from "@chakra-ui/icons";
-import {
-    FaFilePdf,
-    FaFileImage,
-    FaFile,
-} from "react-icons/fa";
 import { parseMessageContent } from "../../../utils/chat/messageParser";
 import { groupActivityTrace } from "../../../utils/chat/activityTrace";
 import ActivityTraceBlock from "../../common/ActivityTraceBlock";
 import MarkdownRenderer from "../../common/MarkdownRenderer";
+import ArtifactCard from "../../common/ArtifactCard";
 import { CitationList } from "../reasoning/components/CitationList";
-
-const formatFileSize = (bytes) => {
-    if (bytes < 1024) return `${bytes} B`;
-    if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-    return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-};
-
-const getArtifactIcon = (mimeType = "") => {
-    if (mimeType.startsWith("image/")) return FaFileImage;
-    if (mimeType === "application/pdf") return FaFilePdf;
-    return FaFile;
-};
-
-const ArtifactCard = ({ artifact }) => {
-    const { filename, mime_type, size, url } = artifact;
-    const Icon = getArtifactIcon(mime_type);
-
-    return (
-        <Box
-            p={2}
-            borderWidth="1px"
-            borderRadius="md"
-            borderColor="gray.200"
-            bg="gray.50"
-            _dark={{ borderColor: "gray.600", bg: "gray.750" }}
-            maxW="320px"
-        >
-            <HStack spacing={2} mb={1}>
-                <Icon size="1.2em" color="gray.500" />
-                <Text fontSize="xs" fontWeight="semibold" isTruncated flex={1}>
-                    {filename}
-                </Text>
-            </HStack>
-            <HStack spacing={2} justify="space-between">
-                <Text fontSize="xs" color="gray.500">
-                    {mime_type} · {formatFileSize(size)}
-                </Text>
-                <HStack spacing={1}>
-                    {mime_type === "application/pdf" && (
-                        <Link href={url} isExternal>
-                            <Button
-                                size="xs"
-                                variant="ghost"
-                                colorScheme="blue"
-                                leftIcon={<ExternalLinkIcon />}
-                                aria-label="View file"
-                            >
-                                View
-                            </Button>
-                        </Link>
-                    )}
-                    <Link href={url} download>
-                        <Button
-                            size="xs"
-                            variant="ghost"
-                            colorScheme="blue"
-                            leftIcon={<DownloadIcon />}
-                            aria-label="Download file"
-                        >
-                            Save
-                        </Button>
-                    </Link>
-                </HStack>
-            </HStack>
-        </Box>
-    );
-};
 
 const ChatMessages = ({
     messages,
